NoahYixiaoHashCode

Documentation for NoahYixiaoHashCode.

Usage

using NoahYixiaoHashCode

problem = load_problem()
solution = solve_greedy(problem)
open("solution_NoahYixiaoHashCode.txt", "w") do f
    save_text(solution; io=f)
end

Optimization Algorithm

Problem and Notations

Suppose we have a city represented by a graph, $G = (\{n\}, \{s\})$. Here, $\{n\} = \{n_i \;\mathrm{where}\; i = 1, 2, \dots, N_n\}$ is a collection of the nodes; $\{s\} = \{s_n \;\mathrm{where}\; n = 1, 2, \dots, N_s\}$ is a collection of the streets. $N_n$ and $N_s$ are the numbers of the nodes and streets respectively. Each street, $s_n$, can be viewed as a tuple, $(x_n, y_n, d_n, t_n, l_n)$. Here, $x_n$ and $y_n$ are the two nodes $s_n$ connect; $d_n$ is whether this street is one-way; $t_n$ is the time cost of traversing $s_n$; $l_n$ is the length of $s_n$.

In this problem, we have multiple cars, the number of which is referred to as $N_c$, They start from the same initial node. The goal is to maximize the total distance of the the streets that all the cars traverse within a limited time.

Greedy algorithm

Our algorithm is a greedy algorithm. In order to avoid traversing certain streets repeatedly, we store how many times each street has been traversed, which will be referred to as $T_n$ where $s_n$ is the $n$-th street. At each step, we maximize the heuristic below among all the streets that the car can traverse the next step.

\[h(s_n) = \left\{\begin{array}{lr}l_n, & T_n = 0 \\ -t_n\times T_n, & T_n > 0 \\ \end{array}\right.\]

If not all the streets have been traversed before, this algorithm maximizes the distance of the next street. If all the streets have been traversed before, this algorithm tries to avoid traversing certain streets over and over again. Actually, $h$ can be any function of $l_n$, $t_n$, and $T_n$. The equation above is just one example that we found easy to calculate and performs well.

Note that $T_n$ is updated at every step. If we decide to traverse $s_n$, we need to do $T_n = T_n + 1$ soon after that. $T_n$ can be larger than $1$.

Look-ahead

We use a look ahead to improve the greedy algorithm. We sum over the values of the heuristic for a certain number of steps. This is done with a tree search and the time complexity is roughly $O(d^L)$, where $d$ is the average degree of nearby vertices and $L$ is the number of steps to look ahead. Note that the change in $T_n$ due to previous steps is considered in the calculation.

Multiple cars

We schedule the routes for all cars sequentially. First, we use the above algorithm for the first car. After running out of time of the first cat, we start to schedule the route of the second car.

However, $T_n$ is shared by all the cars. So, the later cars can also learn from previous cars.

Upper Bound Algorithm

To compute our upper bound, we relax one of the rules related to traveling. Normally, at any timestep, cars on a graph $G$ can only travel along one of the streets adjacent to the junction they're currently at. Instead, we allow cars to travel along any street at any timestep, meaning essentially every junction is connected to every other junction with distance 0 and time cost 0 on a larger graph $G'$. By doing this, we skip the need for good routing since the street taken no longer affects the possible streets to choose from in the next step.

The maximum distance coverable in $G'$ must be at least as much as the maximum distance coverable in $G$ (for the same total amount of time), since $G'$ contains all the edges of $G$ and then some.

To find an upper bound on the maximum distance coverable on $G'$, in at most time $T$, we employ the following algorithm $A$:

Compute the "efficiency" of each street as its length divided by its time cost, and sort all streets in an array from greatest to least efficiency.

Set up running totals for total time and total distance covered, initializing both to 0.

While total time has not yet reached $T$, pop the next street from the front of the array (the street with the highest efficiency), and add its time and distance to the running totals.

Once the total time has met or exceeded $T$, return the total distance.

The total time cost of algorithm $A$ is at least $T$. Any algorithm $B$ that tries to route the cars following the original rules must take at most time $T$. This means in order for algorithm $B$ to cover a greater distance, it would need an average efficiency greater than that of algorithm $A$. However, algorithm $A$ already chooses the maximum possible efficiency, so algorithm $B$ cannot possibly be better.

In conclusion, the algorithm above is a valid upper bound on the maximum possible distance under the original rules. Running the algorithm on the sample Paris data with a maximum time of 18,000 yields an upper bound of 1,525,503.

Index

NoahYixiaoHashCode.Junction

NoahYixiaoHashCode.RoutingProblem

NoahYixiaoHashCode.Solution

NoahYixiaoHashCode.Street

NoahYixiaoHashCode.distance

NoahYixiaoHashCode.empty_solution

NoahYixiaoHashCode.is_feasible

NoahYixiaoHashCode.load_problem

NoahYixiaoHashCode.save_text

NoahYixiaoHashCode.solve_greedy

NoahYixiaoHashCode.time_cost

NoahYixiaoHashCode.total_distance

NoahYixiaoHashCode.total_distance_upper_bound

NoahYixiaoHashCode.Junction — Type

Junction

Store the location of a junction.

Fields

lat::N: latitude

lon::N: longitude

NoahYixiaoHashCode.RoutingProblem — Type

RoutingProblem

Store the information of a routing problem.

Fields

n_junctions::Int: number of junctions

n_cars::Int: number of cars

n_streets::Int: number of streets

total_time::Int: time limit

init_j::Int: id of the initial junction

junctions::J: list of all the Junctions

streets::S: list of all the Streets

sid_matrix::M: a table that shows the indices of streets. It should be a SparseArrays.SparseMatrixCSC, and sid_matrix[j_end, j_begin] is the index of the street if it exists otherwise 0.

NoahYixiaoHashCode.Solution — Type

Solution

Store a solution to a RoutingProblem

Fields

problem::RoutingProblem: the problem to solve

routes: list of junctions that each car visits

NoahYixiaoHashCode.Street — Type

Street

Store the information of a street.

Fields

distance::D: length

time_cost::T: time cost

NoahYixiaoHashCode.distance — Method

distance(street)

Return the length of a Street.

NoahYixiaoHashCode.empty_solution — Method

empty_solution(p::RoutingProblem)

Return an empty solutoin, in which no car moves.

NoahYixiaoHashCode.is_feasible — Method

is_feasible(solution::Solution)

Return whether a Solution is feasible.

NoahYixiaoHashCode.load_problem — Function

load_problem(problem_input=normpath(joinpath(@__DIR__, "..", "data", "paris_54000.txt")))

Return a RoutingProblem defined by the input file.

NoahYixiaoHashCode.save_text — Method

save_text(solution::Solution; io=stdout)

Write the solution in the text format to io.

NoahYixiaoHashCode.solve_greedy — Method

solve_greedy(problem::RoutingProblem; search_depth = 10)

Solve a RoutingProblem using a greedy algorithm.

Arguments

problem: the RoutingProblem to solve

search_depth: the number of steps for the look-ahead. When it is 1, the code uses a greedy algorithm without a look-ahead.

NoahYixiaoHashCode.time_cost — Method

time_cost(street)

Return the time cost of traversing a Street.

NoahYixiaoHashCode.total_distance — Method

total_distance(solution::Solution; check=true)

Return the total distance achieved by a solution.

NoahYixiaoHashCode.total_distance_upper_bound — Method

total_distance_upper_bound(problem::RoutingProblem)

Compute an upper bound on the given RoutingProblem following the algorithm descibed in upper_bound_doc.md