NoahYixiaoHashCode

Documentation for NoahYixiaoHashCode.

NoahYixiaoHashCode.Junction

NoahYixiaoHashCode.RoutingProblem

NoahYixiaoHashCode.Solution

NoahYixiaoHashCode.Street

NoahYixiaoHashCode.distance

NoahYixiaoHashCode.empty_solution

NoahYixiaoHashCode.is_feasible

NoahYixiaoHashCode.load_problem

NoahYixiaoHashCode.save_text

NoahYixiaoHashCode.solve_greedy

NoahYixiaoHashCode.solve_randwalk

NoahYixiaoHashCode.time_cost

NoahYixiaoHashCode.total_distance

NoahYixiaoHashCode.Junction — Type

Junction

Store the location of a junction.

Fields

lat::N: latitude

lon::N: longitude

NoahYixiaoHashCode.RoutingProblem — Type

RoutingProblem

Store the information of a routing problem.

Fields

n_junctions::Int: number of junctions

n_cars::Int: number of cars

n_streets::Int: number of streets

total_time::Int: time limit

init_j::Int: id of the initial junction

junctions::J: list of all the Junctions

streets::S: list of all the Streets

sid_matrix::M: a table that shows the indices of streets

NoahYixiaoHashCode.Solution — Type

Solution

Store a solution to a RoutingProblem

Fields

problem::RoutingProblem: the problem to solve

routes: list of junctions that each car visits

NoahYixiaoHashCode.Street — Type

Street

Store the information of a street.

Fields

distance::D: length

time_cost::T: time cost

NoahYixiaoHashCode.distance — Method

distance(street)

Return the length of a Street.

NoahYixiaoHashCode.empty_solution — Method

empty_solution(p::RoutingProblem)

Return an empty solutoin, in which no car moves.

NoahYixiaoHashCode.is_feasible — Method

is_feasible(solution::Solution)

Return whether a Solution is feasible.

NoahYixiaoHashCode.load_problem — Function

load_problem(problem_input=normpath(joinpath(@__DIR__, "..", "data", "paris_54000.txt")))

Return a [RoutingProblem] defined by the input file.

NoahYixiaoHashCode.save_text — Method

save_text(solution::Solution; io=stdout)

Write the solution in the text format to io.

NoahYixiaoHashCode.solve_greedy — Method

solve_greedy(problem::RoutingProblem)

Solve a RoutingProblem using a greedy algorithm.

NoahYixiaoHashCode.solve_randwalk — Method

solve_randwalk(problem::RoutingProblem)

Solve a RoutingProblem using random walk.

NoahYixiaoHashCode.time_cost — Method

time_cost(street)

Return the time cost of traversing a Street.

NoahYixiaoHashCode.total_distance — Method

total_distance(solution::Solution; check=true)

Return the total distance achieved by a solution.